Add the NIS outlier rejection test to the filter
